feat: Use blueprint 0.58.1 isItemInt flag for integer list items
Number list items now map to int when the blueprint marks them with the
new isItemInt flag, restoring List[int] for supported_code_lengths. All
integer types now match the pre-refactor generated output.

+++ seam/routes/access_codes.py
@@ -233,7 +233,7 @@ class AccessCodes(AbstractAccessCodes):
         device_id: str,
         max_code_length: Optional[int] = None,
         min_code_length: Optional[int] = None,
-        supported_code_lengths: Optional[List[float]] = None
+        supported_code_lengths: Optional[List[int]] = None
     ) -> None:
         json_payload = {}
